package/toolmate/requirements.txt
llama-cpp-python[server]==0.2.90
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
numpy==1.26.4
protobuf==4.25.3
cmake==3.29.3
ollama==0.3.3
groq==0.11.0
openai==1.45.0
tiktoken==0.7.0
google-cloud-aiplatform==1.66.0
pyautogen[retrievechat,autobuild]==0.3.0
autogenstudio==0.1.5
transformers==4.42.3
sentence-transformers==3.0.1
chromadb==0.5.5
soundfile==0.12.1
sounddevice==0.4.7
elevenlabs==1.8.0
huggingface-hub==0.24.7
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.14
langchain-huggingface==0.0.3
guidance==0.1.15
requests
argparse
pendulum
folium
seaborn[stats]
sympy
prompt_toolkit
Pygments
datetime
netifaces
geocoder
googlesearch-python
art
apsw
gTTS
google-cloud-speech
google-cloud-texttospeech
pywhatkit
yt-dlp
rembg
qrcode
pyperclip
colorama
pillow
docker
einops
unstructured[all-docs]
pygame
PySide6
feedparser
html2text
pypdf
PyMuPDF
yfinance
setuptools-rust
SpeechRecognition
openai-whisper
pydub
stable-diffusion-cpp-python
pytz
geopy
psutil